début ben conv

+++ PatriciaTree/src/trie/RTrie.java
@@ -32,7 +32,7 @@ public interface RTrie {
 	
 	public int hauteur();
 	
-	public int profondeurMoyenne();
+	public double profondeurMoyenne();
 	
 	public int prefixe(String mot);
